// front/src/models/productModel.ts
export interface ProductModel {
  _id: string;
  name: string;
  categoryId: string;
  imageUrl: string;
  price: number;
  description: string;
  producer: string;
  createdAt: Date;
  updatedAt: Date;
}

// _id "63dd04a0299c12a5079bba39"
// name "test 1"
// categoryId "63dc68fbcbe985292981c547"
// imageUrl "test1.jpg"
// price 7000
// description "test 1 설명"
// manufacturer "test 1 생산지"
// createdAt
// updatedAt